Arduino connect to web with Python and Flask
This code is to connect arduino with flask via python and pyserial coding.

(Disclaimer: This coding is not mine. This project belongs to poopypigeon from instructables.com. This is to help me push further understanding how to work arduino with flask/python via serial port.)

Tested webpage. Working to expand the code to use it more effective interface for larger interfaces.